+++ BashTerm/GameContext.cs
@@ -1,0 +1,3 @@
+namespace BashTerm;
+
+internal static class GameContext { }