+++ BashTerm/Patch.cs
@@ -8,8 +8,7 @@ internal class Patch {
 	[HarmonyPatch(
 		typeof(LG_ComputerTerminalCommandInterpreter),
 		nameof(LG_ComputerTerminalCommandInterpreter.EvaluateInput)
-		//nameof(LG_ComputerTerminalCommandInterpreter.TryGetCommand)
-		)]
+	)]
 	[HarmonyPrefix]
 	// TODO: Custom aliases defined via config
 	public static void PreEvaluateCmd(ref string inputString) {
@@ -18,7 +17,7 @@ internal class Patch {
 		string param1 = "";
 		string param2 = "";
 
-		if(ConfigMaster.DEBUG) Logger.Info("Pre-Evaluate Command Received = \"" + inputString + "\"");
+		Logger.Debug("Pre-Evaluate Command Received = \"" + inputString + "\"");
 
 		if (args.Length == 0)
 			return;
@@ -28,74 +27,97 @@ internal class Patch {
 		if (args.Length > 2)
 			param2 = args[2];
 
-		switch (cmd) {
-			case "l":
-			case "ls":
-				cmd = "list";
+		TermCmd cmdInternal = Util.InterpretCmd(cmd);
+		bool doTranslate = true;
+
+		switch (cmdInternal) {
+			case TermCmd.ShowList:
+				if (ConfigMaster.LsConvertsNum2ZoneId) {
+					bool firstIsInt = Util.IsInt(param1);
+					bool secondIsInt = Util.IsInt(param2);
+
+					if (firstIsInt) {
+						param1 = "e_" + param1;
+						if (!secondIsInt && ConfigMaster.LsObjExpansions) {
+							param2 = Util.InterpretObject(param2);
+						}
+					} else {
+						if (secondIsInt)
+							param2 = "e_" + param2;
+						if (ConfigMaster.LsObjExpansions)
+							param1 = Util.InterpretObject(param1);
+					}
+				}
 				break;
-			case "lu":
-			case "lsu":
+			case TermCmd.ShowListU:
+				param2 = "";
 				if (param2 == "") {
 					cmd = "list";
 					if (param1 != "" && int.TryParse(param1, out int n)) {
 						param2 = "e_" + param1;
-					} else
-					{
+					} else {
 						param2 = param1;
 					}
 					param1 = "u";
 				}
 				break;
-			case "uc":
-				cmd = "uplink_connect";
+			case TermCmd.TerminalUplinkVerify:
+				doTranslate = ConfigMaster.EnableUVAlias;
 				break;
-			case "rs":
-			case "start":
-				cmd = "reactor_startup";
+			case TermCmd.ReactorVerify:
+				doTranslate = ConfigMaster.EnableRVAlias;
 				break;
-			case "rsd":
-			case "shut":
-			case "shutdown":
-				cmd = "reactor_shutdown";
-				break;
-			case "uv":
-				if (ConfigMaster.EnableUVAlias)
-                    cmd = "uplink_verify";
-				break;
-			case "rv":
-				if (ConfigMaster.EnableRVAlias)
-                    cmd = "reactor_verify";
-				break;
-			case "p":
-			case "ping":
+			case TermCmd.Ping:
 				cmd = "ping";
-				// TODO: This logic is limited, but works for now
-				if (args.Length > 3 || 
-					// below is considering the case they didn't executed a command with -T (e.g. ping ??? -t)
-					(args.Length == 3 && !param1.StartsWith("-") && Util.IsInt(param2))) {
-                    param1 = Util.InterpretObjectType(args[1]) + "_" + Util.Concat('_', args, 2, args.Length);
-                    param2 = "";
-				}				
-				break;
-			case "q":
-			case "query":
-				cmd = "query";
-				if (args.Length > 2) { 
-                    param1 = Util.InterpretObjectType(args[1]) + "_" + Util.Concat('_', args, 2, args.Length);
-                    param2 = "";
+				if (
+					args.Length > 3
+					|| (args.Length == 3 && !Util.ContainsFlag(args) && Util.IsInt(param2))
+				) {
+					param1 =
+						Util.InterpretObject(args[1])
+						+ "_"
+						+ Util.Concat("_", args, 2, args.Length);
+					param2 = "";
 				}
 				break;
-			case "clear":
-				cmd = "cls";
+			case TermCmd.Query:
+				cmd = "query";
+				if (args.Length > 2) {
+					param1 =
+						Util.InterpretObject(args[1])
+						+ "_"
+						+ Util.Concat("_", args, 2, args.Length);
+					param2 = "";
+				}
 				break;
+			case TermCmd.Raw:
+				inputString = Util.Concat(args, 1, args.Length);
+				return;
+				/*
+				case TermCmd.TerminalUplinkConnect:
+					break;
+				case TermCmd.ReactorStartup:
+					break;
+				case TermCmd.ReactorShutdown:
+					break;
+				case TermCmd.ReadLog:
+					break;
+				case TermCmd.Cls:
+					break;
+				*/
 		}
 
-		if (cmd == "raw" || cmd == "r") {
-			inputString = Util.Concat(args, 1, args.Length);
-		} else {
-            inputString = Util.Concat(cmd, param1, param2).ToUpper().Trim();
+		Logger.Debug("We interpreted cmdInternal = " + cmdInternal.ToString("G"));
+
+		if (cmdInternal != TermCmd.None && doTranslate) {
+			cmd = Util.Cmd2Txt(cmdInternal);
+			Logger.Debug("Command is not None and we are translating, yay! to: " + cmd);
 		}
 
-		if(ConfigMaster.DEBUG) Logger.Info("Command Evaluated as \"" + inputString + "\"");
+		inputString = Util.Concat(cmd, param1, param2);
+
+		inputString = inputString.ToUpper().Trim();
+
+		Logger.Debug("Command Evaluated as \"" + inputString + "\"");
 	}
 }