+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# Change Log
+
+## 0.2.0
+
+* Add additional aliases
+  * `CAT` to `READ`
+* Add additional expansions
+  * `HIS`+ to `HISEC_CARGO`
+  * `NFR`+ to `NFRAME`
+* README
+  * Add description on `RAW` input
+
 ## 0.1.0
 
 - Initial release